Features and fixes
added copy and meal prep functions. mp allows you to multiply the ingredients based on a proven index. copy allows you to copy believe it or not.

fixed a few css errors as well as include new code for the features. Addeda snackbar animation for the copying.

fixed a bug regarding the page not scrolling back to top after home or recipe buttons are pressed.